requirements.txt
albumentations==1.3.0
dependency-injector==4.41.0
pybind11
fire==0.5.0
fastwer==0.1.3  
Levenshtein==0.20.8
lightning-utilities==0.8.0
lxml==4.9.2 
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.2
pytorch-lightning==2.0.2
Pillow==9.3.0 
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.9.3
tensorflow==2.13.0
timm==0.5.4
torch
torchaudio
torchmetrics
torchvision
tqdm
transformers==4.26.0
trdg==1.8.0
wandb==0.14.0
